libinput: set_calibration_matrix bandaid
If set to the identity matrix, libinput does not store the calibration
matrix so libinput_device_config_calibration_get_matrix will give the
last non-identity matrix set. This makes it so we cannot reliably test
whether the matrix has changed.

This is just a bandaid fix that will unconditionally set the
calibration matrix. Also, since we cannot reliably determine whether
the matrix has changed, `set_calibration_matrix` will return true
resulting in the `input::libinput_config` ipc event being sent whenever
it is called for a device that supports calibration matrices. If the
input config does not have a calibration matrix configured or the
device does not support calibration matrices, then this bandaid will
have no change.

A libinput MR has been submitted to properly fix the issue. When it
makes it to a libinput release, this commit should be reverted.